+++ RELEASE_NOTES.md
@@ -20,8 +20,10 @@ Older versions are kept in the `Release_Notes` folder.
 
 ### ✨ New features
 
-* `LocaleDirectionProvider` is a new protocol that exposes `Locale` extensions to the DocC documentation. 
-* `LocaleNameProvider` is a new protocol that exposes `Locale` extensions to the DocC documentation. 
+* `KeyboardLocale` has new `sorted` functions.
+* `Locale` has new `localizedName` and `sorted` functions.
+* `LocaleDirectionProvider` is a new protocol that exposes `Locale` extensions to DocC. 
+* `LocaleNameProvider` is a new protocol that exposes `Locale` name extensions to DocC. 
 
 ### 💡 Behavior changes